requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
beautifulsoup4==4.6.3
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
daphne==2.2.4
Django==2.1.4
django-dynamic-preferences==1.7
hiredis==0.3.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
langdetect==1.0.7
metapy==0.2.13
msgpack==0.5.6
msgpack-python==0.5.6
nltk==3.4
numpy==1.15.4
pandas==0.23.4
persisting-theory==0.2.1
PyHamcrest==1.9.0
python-dateutil==2.7.5
pytz==2018.7
redis==2.10.6
requests==2.21.0
scipy==1.1.0
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
websockets==7.0
zope.interface==4.6.0
